app/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]

build-backend = "setuptools.build_meta"
[tool.setuptools.packages]
find = {}

[project]
name = "isep-ctf-backend"
version = "0.0.1"
description = "Backend module utilizing fastAPI"
readme = "README.md"
requires-python = ">=3.10"
keywords = []
authors = [
    { "name"="Jakub Sukdol", "email"="jakub.sukdol@devbal.com" },# TODO: add other authors
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
]
dependencies = [
    "charset-normalizer>=3.4.3",
    "fastapi>=0.118.0",
    "passlib>=1.7.4",
    "pre-commit>=4.3.0",
    "pydantic[email]>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.43",
    "psycopg[binary]>=3.1",
    "asyncpg>=0.29.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
    "argon2-cffi>=25.1.0",
    "alembic>=1.17.1",
    "aiosqlite>=0.21.0",
    "typer>=0.20.0",
    "redis>=5.0",
    "Pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "loguru>=0.7.3",
    "slowapi>=0.1.9",
    "psycopg2-binary>=2.9.11",
    "pyotp>=2.9.0",
]


[tool.setuptools]
include-package-data = true

[tool.mypy]
python_version = "3.10"
# Exclude Alembic migration files from type-checking
exclude = "^app/backend/alembic/.*"
# Avoid duplicate-module-name errors when running mypy from the repo root
explicit_package_bases = true

# Silence missing-stub errors for SQLAlchemy (install types if you want stricter checks)
[tool.mypy."sqlalchemy.*"]
ignore_missing_imports = true

# Ignore all Alembic modules entirely
[tool.mypy."app.backend.alembic.*"]
ignore_errors = true
